subdivision into more components
